Generate a classification report in a pandas DataFrame format
This small madule is created to simply convert a classification report into a pandas dataframe for the ease of use and comparision. It will also be easier to use to_csv method to send the resuls to csv/excel and put them on a presentation slide.
Feel free to copy/edit the code for any purpose bute take the code as is, also no guarantees of any sort, but please report me of any possible errors.